relocate core files (improve use as library)
