+++ localrun.py
@@ -1,4 +1,4 @@
-from src.lanscape.main import main
+from src.lanscape.libraries.core.main import main
 
 if __name__ == "__main__":
     main()
