+++ src/lanscape/__main__.py
@@ -1,4 +1,6 @@
-from .main import main
+from .libraries.core.main import main
+
+# leveraged when calling as module ie "python -m lanscape"
 
 if __name__ == "__main__":
     main()
